+++ 3days/app/src/main/java/com/weave/a3days/DaysNavGraph.kt
@@ -5,9 +5,9 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import com.weave.home.navGraphHome
 import com.weave.intro.navGraphIntro
 import com.weave.my_profile.navGraphMyProfile
-import com.weave.navGraphHome
 import com.weave.utils.LoggerUtil
 
 enum class Route(val routeName: String) {
